examples/boltzmann: Restructure
- Restructure into agents.py and model.py
- Make compute_gini a Model function
