--- examples/basic/boltzmann_wealth_model/requirements.txt
@@ -1,1 +1,0 @@
-mesa[viz]>=3.0.0b0
